infra/requirements.txt
aws-cdk-lib>=2.243.0
constructs>=10.0.0
aws-cdk.aws-bedrock-agentcore-alpha>=2.243.0a0
